+++ CHANGELOG.md
@@ -58,4 +58,8 @@ All notable changes to `datum` will be documented in this file
 
 ## 0.9.0 - 2021-02-11
 - add abstract queryFilters() method to ApplyFilter for declaring an array of Filters 
-- cut $attribute_filters properties from AbstractQueryWithFilter(s) and replaced with queryFilters() methods._
+- cut $attribute_filters properties from AbstractQueryWithFilter(s) and replaced with queryFilters() methods.
+
+
+## 0.9.1 - 2021-02-11
+- make Sfneal\Queries\Traits\HasRelationships trait for dynamically adding relationships to a query